# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Summer School Community Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register with invite code
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: Registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current session user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ── Profiles ──────────────────────────────────────────────────────────────
...
  /posts/{postId}/comments:
    get:
      operationId: listComments
      tags: [comments]
      summary: List comments on a post
      parameters:
        - name: postId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Comments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Comment"

    post:
      operationId: createComment
      tags: [comments]
      summary: Add a comment to a post
      parameters:
        - name: postId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CommentInput"
      responses:
        "201":
          description: Created comment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Comment"
  /comments/{id}:
    delete:
      operationId: deleteComment
      tags: [comments]
      summary: Delete a comment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── Events ────────────────────────────────────────────────────────────────
  /events:
    get:
      operationId: listEvents
      tags: [events]
      summary: List events
      parameters:
        - name: upcoming
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: List of events
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Event"

    post:
      operationId: createEvent
      tags: [events]
      summary: Create an event (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EventInput"
      responses:
        "201":
          description: Created event
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
  /events/{id}:
    get:
      operationId: getEvent
      tags: [events]
      summary: Get an event by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Event
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    patch:
      operationId: updateEvent
      tags: [events]
      summary: Update an event (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EventUpdate"
      responses:
        "200":
          description: Updated event
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"

    delete:
      operationId: deleteEvent
      tags: [events]
      summary: Delete an event (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /events/{id}/rsvp:
    post:
      operationId: rsvpEvent
      tags: [events]
      summary: RSVP to an event
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RsvpInput"
      responses:
        "200":
          description: RSVP recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
  /events/{id}/rsvps:
    get:
      operationId: getEventRsvps
      tags: [events]
      summary: Get RSVPs for an event (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: RSVP list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Rsvp"

  # ── Dashboard (wow endpoints) ──────────────────────────────────────────────
...
  /admin/students/{id}/deactivate:
    patch:
      operationId: adminDeactivateStudent
      tags: [admin]
      summary: Deactivate a student account (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminStudent"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    RegisterInput:
      type: object
      required: [email, password, inviteCode]
      properties:
        email:
          type: string
        password:
          type: string
        inviteCode:
          type: string
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
    AuthResponse:
      type: object
      required: [user]
      properties:
        user:
          $ref: "#/components/schemas/User"
    User:
      type: object
      required: [id, email, role, onboardingComplete]
      properties:
        id:
          type: integer
        email:
          type: string
        role:
          type: string
          enum: [student, admin]
        onboardingComplete:
          type: boolean

    # ── Profile ─────────────────────────────────────────────────────────────
...
    Comment:
      type: object
      required: [id, postId, authorId, authorName, content, createdAt]
      properties:
        id:
          type: integer
        postId:
          type: integer
        authorId:
          type: integer
        authorName:
          type: string
        authorPhoto:
          type: ["string", "null"]
        content:
          type: string
        parentCommentId:
          type: ["integer", "null"]
        createdAt:
          type: string
    CommentInput:
      type: object
      required: [content]
      properties:
        content:
          type: string
        parentCommentId:
          type: integer

    # ── Events ──────────────────────────────────────────────────────────────
    Event:
      type: object
      required: [id, title, description, eventType, startAt, endAt, location, status, rsvpCounts, myRsvp, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        eventType:
          type: string
          enum: [workshop, social, guestTalk, fieldTrip, other]
        startAt:
          type: string
        endAt:
          type: string
        timezone:
          type: string
        location:
          type: string
        meetingUrl:
          type: ["string", "null"]
        coverImageUrl:
          type: ["string", "null"]
        capacity:
          type: ["integer", "null"]
        rsvpDeadline:
          type: ["string", "null"]
        status:
          type: string
          enum: [draft, published, cancelled, past]
        rsvpCounts:
          $ref: "#/components/schemas/RsvpCounts"
        myRsvp:
          type: ["string", "null"]
        createdAt:
          type: string
    EventInput:
      type: object
      required: [title, description, eventType, startAt, endAt, location]
      properties:
        title:
          type: string
        description:
          type: string
        eventType:
          type: string
          enum: [workshop, social, guestTalk, fieldTrip, other]
        startAt:
          type: string
        endAt:
          type: string
        timezone:
          type: string
        location:
          type: string
        meetingUrl:
          type: string
        coverImageUrl:
          type: string
        capacity:
          type: integer
        rsvpDeadline:
          type: string
        status:
          type: string
          enum: [draft, published, cancelled, past]
    EventUpdate:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        eventType:
          type: string
        startAt:
          type: string
        endAt:
          type: string
        timezone:
          type: string
        location:
          type: string
        meetingUrl:
          type: string
        coverImageUrl:
          type: string
        capacity:
          type: integer
        rsvpDeadline:
          type: string
        status:
          type: string
          enum: [draft, published, cancelled, past]
    RsvpCounts:
      type: object
      properties:
        going:
          type: integer
        notGoing:
          type: integer
        maybe:
          type: integer
    RsvpInput:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [going, notGoing, maybe]
    Rsvp:
      type: object
      required: [userId, studentName, status]
      properties:
        userId:
          type: integer
        studentName:
          type: string
        studentCollege:
          type: ["string", "null"]
        photoUrl:
          type: ["string", "null"]
        status:
          type: string

    # ── Dashboard ────────────────────────────────────────────────────────────
...
    AdminStudent:
      type: object
      required: [id, email, role, isActive, onboardingComplete, joinedAt]
      properties:
        id:
          type: integer
        email:
          type: string
        role:
          type: string
        isActive:
          type: boolean
        onboardingComplete:
          type: boolean
        profileName:
          type: ["string", "null"]
        college:
          type: ["string", "null"]
        joinedAt:
          type: string
...