# lib/api-zod/src/generated/types/adminStudent.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Summer School Community Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface AdminStudent {
  id: number;
  email: string;
  role: string;
  isActive: boolean;
  onboardingComplete: boolean;
  /** @nullable */
  profileName?: string | null;
  /** @nullable */
  college?: string | null;
  joinedAt: string;
}