# lib/api-zod/src/generated/types/rsvp.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Summer School Community Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface Rsvp {
  userId: number;
  studentName: string;
  /** @nullable */
  studentCollege?: string | null;
  /** @nullable */
  photoUrl?: string | null;
  status: string;
}